Refactor all tests to use pytest
Also skip Python 3.4 tests as it was failing on
TypeError: skip() got an unexpected keyword argument 'allow_module_level'
