+++ .travis.yml
@@ -6,9 +6,6 @@ matrix:
       python: "2.7"
       env: TOXENV=py27
     - os: linux
-      python: "3.4"
-      env: TOXENV=py34
-    - os: linux
       python: "3.5"
       env: TOXENV=py35
     - os: linux
